[#51] Remove unused type, type assertions
